feat: complete XML-based version with full CRUD, ManyToMany, navbar UI and realistic data initialization

+++ src/main/java/com/ximena/discografia/repository/TrackRepository.java
@@ -9,7 +9,7 @@ public interface TrackRepository {
 
     void save(Track track);
 
-    void deleteById(Long id);
+    void deleteById(int id);
 
-    Track findById(Long id);
+    Track findById(int id);
 }
